+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-relay-offline",
-  "version": "0.1.4-a15",
+  "version": "0.2.0",
   "keywords": [
     "graphql",
     "relay",
